+++ metaconfigure/description.json
@@ -385,8 +385,8 @@
     "src/ACEtk/interpretation/EL03/RileyCrossSection.hpp",
     "src/ACEtk/interpretation/EPRData/Coherent.hpp",
     "src/ACEtk/interpretation/EPRData/CRTP.hpp",
-    "src/ACEtk/interpretation/EPRData/FormFactors.hpp",
     "src/ACEtk/interpretation/EPRData/Incoherent.hpp",
+    "src/ACEtk/interpretation/EPRData/CRTP/FormFactors.hpp",
     "src/ACEtk/interpretation/MP1/DiscreteData.hpp",
     "src/ACEtk/interpretation/MP1/src/cummulativeDistribution.hpp",
     "src/ACEtk/interpretation/MP1/src/energyGrid.hpp",
